requirements.txt
Appium-Python-Client==2.11.1
allure-pytest==2.13.2
requests==2.24.0
selenium==4.10.0
pytest==7.3.2
